+++ pkg/R/lefko3-package.R
@@ -48,6 +48,7 @@
 #' @importFrom Matrix colSums
 #' @importFrom methods as is
 #' @importFrom MuMIn dredge
+#' @importFrom pscl zeroinfl
 #' @importFrom Rcpp evalCpp
 #' @importFrom stats getCall glm lm na.action na.fail na.omit rnorm sd setNames xtabs
 #' @importFrom stats as.formula median pchisq poisson var logLik complete.cases